requirements.txt
flask
pymongo
google-generativeai
rapidfuzz
flask-cors
gunicorn
